requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.2.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.110.1
h11==0.14.0
httptools==0.6.1
idna==3.6
packaging==24.0
passlib==1.7.4
psutil==5.9.8
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
sniffio==1.3.1
sse-starlette==2.1.0
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
